+++ CHANGELOG.md
@@ -24,9 +24,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Adds `--playground-code-padding` and `--playground-code-line-padding` for
   configuring code editor padding.
 
+- Bare module specifiers are now transformed in dynamic `import()` statements.
+
 ### Changed
 
 - **BREAKING** The `src/` directory is no longer published to NPM.
+
+- **BREAKING** `.js` files are no longer compiled by TypeScript, so they cannot
+  contain types, decorators, or other TypeScript-specific syntax.
 
 - Previews will now begin loading immediately, instead of waiting for
   compilation to completely finish, and each `.js` file is served as it